app/worker/requirements.txt
azure-identity>=1.16.0
azure-servicebus~=7.12
opencensus-ext-azure==1.1.13
azure-data-tables>=12.6.0
requests>=2.32.3
